api/requirements.txt
fastapi[standard]==0.115.0
google-generativeai==0.8.1
python-dotenv==1.0.1
